+++ test.sh
@@ -3,22 +3,23 @@
 # Test script for https://github.com/KazKobara/dockerfile_fswiki_local .
 #
 
-## Edit here
+### To test both Alpine and Debian/Ubuntu or either of them.
+## Uncomment one of them
 TEST_PLATFORM="alpine ubuntu"
 # TEST_PLATFORM="alpine"
 # TEST_PLATFORM="ubuntu"
 
 ## Comment out if not to test
-TEST_DOCKER_COMPOSE="Do"
-TEST_SHELL_VERSION="Do"
+TEST_COMPOSE_VER="Do"
+TEST_SHELL_VER="Do"
 
 TEST_TMP=.test.tmp
 
 ### check ###
 
-if [ -z "${TEST_DOCKER_COMPOSE+x}" ] && [ -z "${TEST_SHELL_VERSION+x}" ]; then  # if not defiend
+if [ -z "${TEST_COMPOSE_VER+x}" ] && [ -z "${TEST_SHELL_VER+x}" ]; then  # if not defiend
 	echo
-	echo " Error: either 'TEST_DOCKER_COMPOSE' or 'TEST_SHELL_VERSION' should be defined!!"
+	echo " Error: either 'TEST_COMPOSE_VER' or 'TEST_SHELL_VER' should be defined!!"
 	exit 1
 fi
 
@@ -28,18 +29,18 @@ fi
 # @param $1 $container_name
 # @param $2 $image_name
 check_and_remove () {
-    if docker inspect --type=container "$1" > /dev/null 2>&1; then
-		docker stop "$1" && docker rm "$1"
+    if ${CONTAINER_CLI} inspect --type=container "$1" > /dev/null 2>&1; then
+		${CONTAINER_CLI} stop "$1" && ${CONTAINER_CLI} rm "$1"
     fi
-    if docker inspect --type=image "$2" > /dev/null 2>&1; then
-		docker rmi "$2"
+    if ${CONTAINER_CLI} inspect --type=image "$2" > /dev/null 2>&1; then
+		${CONTAINER_CLI} rmi "$2"
     fi
 }
 
 # @brief Check `wget http://"${ADDR_AND_PORT}"/wiki.cgi``.
 # @exit 1 if failed.
 check_wget () {
-    source .env
+    # source .env
     if echo "${FSWIKI_PORT}" | grep ':'; then
 		# addr:port
 		ADDR_AND_PORT="${FSWIKI_PORT}"
@@ -54,20 +55,24 @@ check_wget () {
 }
 
 ### main ###
-
-DOCKER_COMPOSE=""
-if [ -n "${TEST_DOCKER_COMPOSE+x}" ]; then  # if defiend
-	if command -v docker-compose; then
-		DOCKER_COMPOSE="docker-compose"
-	else
-		if command -v docker-compose.exe; then
-			DOCKER_COMPOSE="docker-compose.exe"
+source .env
+if [ "${COMPOSE}" == "docker-compose" ]; then
+	DOCKER_COMPOSE=""
+	if [ -n "${TEST_COMPOSE_VER+x}" ]; then  # if defined
+		if command -v docker-compose; then
+			DOCKER_COMPOSE="docker-compose"
 		else
-			echo
-			echo "WARNING: no docker-compose command found and skip them!!"
-			echo
+			if command -v docker-compose.exe; then
+				DOCKER_COMPOSE="docker-compose.exe"
+			else
+				echo
+				echo "WARNING: no docker-compose command found and skip them!!"
+				echo
+			fi
 		fi
 	fi
+else
+	DOCKER_COMPOSE="${COMPOSE}"
 fi
 
 cp -pf .env .env.org
@@ -83,7 +88,7 @@ for f_platform in ${TEST_PLATFORM} ;do
 		sed -i "/^FSWIKI_VERSION=/cFSWIKI_VERSION=$f_version" .env
 		if [ "$DOCKER_COMPOSE" != "" ]; then
 	    	echo
-		    echo "=== docker-compose version ==="
+		    echo "=== Compose version ==="
 		    container_pre_name=fswiki_${f_platform}_local_dc
 		    container_name=${container_pre_name}_${f_version}
 	    	image_name=$container_pre_name:$f_version
@@ -92,12 +97,13 @@ for f_platform in ${TEST_PLATFORM} ;do
 	    	$DOCKER_COMPOSE up -d
 		    echo
 		    echo "=== Versions of $image_name ===$f_ver_rem"         | tee -a $TEST_TMP
-			./check_versions.sh "$container_name"                    | tee -a $TEST_TMP
+			# ./check_versions.sh "$container_name"                    | tee -a $TEST_TMP
+			./check_ver_in_container.sh "$container_name"            | tee -a $TEST_TMP
 		    echo "-------------------------------------------------" | tee -a $TEST_TMP
 		    check_wget
 	    	$DOCKER_COMPOSE down
 		fi
-		if [ -n "${TEST_SHELL_VERSION+x}" ]; then  # if defiend
+		if [ -n "${TEST_SHELL_VER+x}" ]; then  # if defined
 			echo
 			echo "=== shell version ==="
 			container_pre_name=fswiki_${f_platform}_local
@@ -108,10 +114,11 @@ for f_platform in ${TEST_PLATFORM} ;do
 			./run_fswiki_local.sh
 			echo
 			echo "=== Versions of $image_name ======$f_ver_rem"      | tee -a $TEST_TMP
-			./check_versions.sh "$container_name"                    | tee -a $TEST_TMP
+			# ./check_versions.sh "$container_name"                    | tee -a $TEST_TMP
+			./check_ver_in_container.sh "$container_name"            | tee -a $TEST_TMP
 			echo "-------------------------------------------------" | tee -a $TEST_TMP
 			check_wget
-			docker stop "$container_name" && docker rm "$container_name"
+			${CONTAINER_CLI} stop "$container_name" && ${CONTAINER_CLI} rm "$container_name"
 		fi
     done
 done
@@ -122,7 +129,7 @@ echo
 echo
 echo "============ Test Summary ($(git log -1 --pretty=%h)) ============="
 echo "====== Size (of all the 'fswiki_*' images) ======"
-docker images | grep fswiki_
+${CONTAINER_CLI} images | grep fswiki_
 echo
 cat $TEST_TMP
 rm -f $TEST_TMP